requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
yt-dlp>=2023.7.6
python-multipart>=0.0.5
python-dotenv>=0.19.0
pydantic>=1.8.0
